refactor: extract mapServerMessages to ChatServerMessageMapper (verbatim move)
